requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2.8
django-extensions==2.2.5
django-reverse-admin==2.7.3
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.8
Markdown==3.1.1
martor==1.4.4
Pillow==6.2.1
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
pyjwt==1.7.1